fix(resource): remove redundant float cast toDp

+++ app/src/main/java/io/github/kdetard/koki/misc/ResourceUtils.java
@@ -8,6 +8,6 @@ public class ResourceUtils {
     }
 
     public static float toDp(float px) {
-        return (float) (px * Resources.getSystem().getDisplayMetrics().density + 0.5f);
+        return px * Resources.getSystem().getDisplayMetrics().density + 0.5f;
     }
 }
